* Added cp_trusted table. * Added DonationTrustTable configuration directive.
The donation trust table (cp_trusted) will be used to provide the ability to credit accounts for a donation based on their trust status, if they are trusted donations will be credited immediately, if not the crediting process will be on hold for a specified duration of time.  This is helpful to prevent fraudulent donations (stolen credit cards, stolen PayPal accounts, chargebacks etc).